Move dendro_loader inside data_factories directory
